backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Bacheca Backend API"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django>=6.0.2",
    "djangorestframework>=3.16.1",
    "django-filter>=25.1",
    "psycopg[binary]>=3.2",
    "python-decouple>=3.8",
    "gunicorn>=23.0",
    "django-cors-headers>=4.6",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-django>=4.9", "pytest-cov>=6.0"]

[dependency-groups]
dev = ["ruff>=0.15.1"]
